moving module function to components directory
